requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
sqlalchemy==2.0.28
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pytest==8.0.2
pytest-playwright==0.4.4
httpx==0.27.0
requests==2.31.0
